New Client.java
i've changed our original code so that it looks pretty and clean as possible. but yea.. i've tested most of the important stuff i made in this code. like file creation and writing to file/ reading from file... yet we need to test it with the server and hopefully there will be no errors.